--- a/crates/wesl-macros/Cargo.toml
+++ b/crates/wesl-macros/Cargo.toml
@@ -13,8 +13,8 @@ proc-macro = true
 
 [dependencies]
 itertools = "0.14.0"
-proc-macro2 = "1.0.86"
 proc-macro-error2 = { version = "2.0.1", optional = true }
+proc-macro2 = "1.0.86"
 quote = "1.0.37"
 syn = "2.0.77"
 token_stream_flatten = { version = "0.1.0", optional = true }