--- a/crates/wgsl-types/Cargo.toml
+++ b/crates/wgsl-types/Cargo.toml
@@ -7,12 +7,10 @@ repository.workspace = true
 license.workspace = true
 
 [dependencies]
-derive_more = { version = "2.0.1", features = ["from", "unwrap"] }
 half = { version = "2.4.1", features = ["num-traits"] }
 itertools = "0.14.0"
 num-traits = "0.2.19"
 serde = { version = "1.0.215", features = ["derive", "rc"], optional = true }
-thiserror = "2.0.11"
 tokrepr = { workspace = true, optional = true }
 
 [features]