--- a/crates/wgsl-types/src/error.rs
+++ b/crates/wgsl-types/src/error.rs
@@ -1,5 +1,3 @@
-use thiserror::Error;
-
 use crate::{
     CallSignature, Type,
     inst::LiteralInstance,
@@ -7,78 +5,120 @@ use crate::{
 };
 
 /// The global error struct.
-#[derive(Clone, Debug, Error)]
+#[derive(Clone, Debug)]
 pub enum Error {
-    #[error("not implemented: `{0}`")]
     Todo(String),
 
     // types & templates
-    #[error("expected a scalar type, got `{0}`")]
     NotScalar(Type),
-    #[error("`{0}` is not constructible")]
     NotConstructible(Type),
-    #[error("invalid sampled type, expected `i32`, `u32` of `f32`, got `{0}`")]
     SampledType(Type),
 
     // references
-    #[error("cannot write a `{0}` to a reference to `{1}`")]
     WriteRefType(Type, Type),
-    #[error("attempt to write to a read-only reference")]
     NotWrite,
-    #[error("attempt to read a write-only reference")]
     NotRead,
-    #[error("reference is not read-write")]
     NotReadWrite,
-    #[error("cannot create a pointer in `handle` address space")]
     PtrHandle,
-    #[error("cannot create a pointer to a vector component")]
     PtrVecComp,
 
     // conversions
-    #[error("cannot convert from `{0}` to `{1}`")]
     Conversion(Type, Type),
-    #[error("overflow while converting `{0}` to `{1}`")]
     ConvOverflow(LiteralInstance, Type),
 
     // indexing
-    #[error("`{0}` has no component `{1}`")]
     Component(Type, String),
-    #[error("`{0}` cannot be indexed")]
     NotIndexable(Type),
-    #[error("index `{0}` is out-of-bounds for `{1}` of `{2}` components")]
     OutOfBounds(usize, Type, usize),
 
     // arithmetic
-    #[error("cannot use unary operator `{0}` on type `{1}`")]
     Unary(UnaryOperator, Type),
-    #[error("cannot use binary operator `{0}` with operands `{1}` and `{2}`")]
     Binary(BinaryOperator, Type, Type),
-    #[error("cannot apply component-wise binary operation on operands `{0}` and `{1}`")]
     CompwiseBinary(Type, Type),
-    #[error("attempt to add with overflow")]
     AddOverflow,
-    #[error("attempt to subtract with overflow")]
     SubOverflow,
-    #[error("attempt to multiply with overflow")]
     MulOverflow,
-    #[error("attempt to divide by zero")]
     DivByZero,
-    #[error("attempt to calculate the remainder with a divisor of zero")]
     RemZeroDiv,
-    #[error("attempt to shift left by `{0}`, which would overflow `{1}`")]
     ShlOverflow(u32, LiteralInstance),
-    #[error("attempt to shift right by `{0}`, which would overflow `{1}`")]
     ShrOverflow(u32, LiteralInstance),
 
     // functions
-    #[error("invalid function call signature: `{0}`")]
     Signature(CallSignature),
-    #[error("{0}")]
     Builtin(&'static str),
-    #[error("invalid template arguments to `{0}`")]
     TemplateArgs(&'static str),
-    #[error("incorrect number of arguments to `{0}`, expected `{1}`, got `{2}`")]
     ParamCount(String, usize, usize),
-    #[error("invalid parameter type, expected `{0}`, got `{1}`")]
     ParamType(Type, Type),
+}
+
+impl std::fmt::Display for Error {
+    fn fmt(&self, fmt: &mut std::fmt::Formatter) -> std::fmt::Result {
+        match self {
+            Error::Todo(v) => write!(fmt, "not implemented: `{v}`"),
+            Error::NotScalar(ty) => write!(fmt, "expected a scalar type, got `{ty}`"),
+            Error::NotConstructible(ty) => write!(fmt, "`{ty}` is not constructible"),
+            Error::SampledType(ty) => write!(
+                fmt,
+                "invalid sampled type, expected `i32`, `u32` or `f32`, got `{ty}`"
+            ),
+            Error::WriteRefType(new_ty, ty) => {
+                write!(fmt, "cannot write a `{new_ty}` to a reference to `{ty}`")
+            }
+            Error::NotWrite => write!(fmt, "attempt to write to a read-only reference"),
+            Error::NotRead => write!(fmt, "attempt to read a write-only reference"),
+            Error::NotReadWrite => write!(fmt, "reference is not read-write"),
+            Error::PtrHandle => write!(fmt, "cannot create a pointer in `handle` address space"),
+            Error::PtrVecComp => write!(fmt, "cannot create a pointer to a vector component"),
+            Error::Conversion(from_ty, to_ty) => {
+                write!(fmt, "cannot convert from `{from_ty}` to `{to_ty}`")
+            }
+            Error::ConvOverflow(literal, ty) => {
+                write!(fmt, "overflow while converting `{literal}` to `{ty}`")
+            }
+            Error::Component(ty, name) => write!(fmt, "`{ty}` has no component `{name}`"),
+            Error::NotIndexable(ty) => write!(fmt, "`{ty}` cannot be indexed"),
+            Error::OutOfBounds(index, ty, num_components) => write!(
+                fmt,
+                "index `{index}` is out-of-bounds for `{ty}` of `{num_components}` components"
+            ),
+            Error::Unary(op, ty) => write!(fmt, "cannot use unary operator `{op}` on type `{ty}`"),
+            Error::Binary(op, left_ty, right_ty) => write!(
+                fmt,
+                "cannot use binary operator `{op}` with operands `{left_ty}` and `{right_ty}`"
+            ),
+            Error::CompwiseBinary(ty_1, ty_2) => write!(
+                fmt,
+                "cannot apply component-wise binary operation on operands `{ty_1}` and `{ty_2}`"
+            ),
+            Error::AddOverflow => write!(fmt, "attempt to add with overflow"),
+            Error::SubOverflow => write!(fmt, "attempt to subtract with overflow"),
+            Error::MulOverflow => write!(fmt, "attempt to multiply with overflow"),
+            Error::DivByZero => write!(fmt, "attempt to divide by zero"),
+            Error::RemZeroDiv => write!(
+                fmt,
+                "attempt to calculate the remainder with a divisor of zero"
+            ),
+            Error::ShlOverflow(num, ty) => write!(
+                fmt,
+                "attempt to shift left by `{num}`, which would overflow `{ty}`"
+            ),
+            Error::ShrOverflow(num, ty) => write!(
+                fmt,
+                "attempt to shift right by `{num}`, which would overflow `{ty}`"
+            ),
+            Error::Signature(call_signature) => {
+                write!(fmt, "invalid function call signature: `{call_signature}`")
+            }
+            Error::Builtin(name) => write!(fmt, "{name}"),
+            Error::TemplateArgs(name) => write!(fmt, "invalid template arguments to `{name}`"),
+            Error::ParamCount(name, expected_count, actual_count) => write!(
+                fmt,
+                "incorrect number of arguments to `{name}`, expected `{expected_count}`, got `{actual_count}`"
+            ),
+            Error::ParamType(expected_ty, actual_ty) => write!(
+                fmt,
+                "invalid parameter type, expected `{expected_ty}`, got `{actual_ty}`"
+            ),
+        }
+    }
 }