This finishes the work of removing derive_more. And then I went ahead and also replaced thiserror with a manual implementation, [as burntsushi recommends](https://www.reddit.com/r/rust/comments/1cnhy7d/comment/l380bpr/).

The incremental compile times are practically unchanged. 

The from scratch compile times used to be ~7.30 seconds, and now they're down to ~4 seconds on my machine. I think that's worth it.

We could get the compile times to be even lower. Current problems are

- half-rs with the num-traits feature unconditionally enables the `libm` feature https://github.com/VoidStarKat/half-rs/commit/661708e91197e7e51e9f95b0f80cf6c2d998f669 . I don't think it should do that.
- itertools takes a while to compile. But wgsl-analyzer [also depends on itertools](https://github.com/search?q=repo%3Awgsl-analyzer%2Fwgsl-analyzer%20itertools&type=code), so I'm fine with it.
  - wesl-rs uses itertools for `collect_vec`, `to_vec`, `format`, `all_equal` and a few one-off things
- num-traits has a really slow build script. It has a serial compilation pipeline with "autocfg -> build.rs of num-traits -> running the build.rs -> num-traits", which takes almost 2 seconds.
- wgsl-types is surprisingly slow to compile. 

<img width="1692" height="377" alt="image" src="https://github.com/user-attachments/assets/0a406517-29b1-4093-8b8d-1beaf958bd05" />


Regarding the wgsl-types compilation times, I ran the following. But I'm not sure if there's anything actionable there.
```
cd crates/wgsl-types                                                                                                                                                                                                                                                    
cargo rustc -- -Zself-profile
cargo install --git https://github.com/rust-lang/measureme summarize
summarize summarize ../../wgsl_types-0013868.mm_profdata                                                                                                                                                                                                 
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| Item                                                                    | Self time | % of total time | Time     | Item count | Incremental result hashing time |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| LLVM_module_codegen_emit_obj                                            | 2.45s     | 43.137          | 2.45s    | 124        | 0.00ns                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| LLVM_passes                                                             | 719.13ms  | 12.661          | 719.13ms | 1          | 0.00ns                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| codegen_module                                                          | 637.85ms  | 11.230          | 748.43ms | 124        | 0.00ns                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| typeck                                                                  | 252.22ms  | 4.441           | 311.13ms | 969        | 7.80ms                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| mir_borrowck                                                            | 203.57ms  | 3.584           | 289.87ms | 707        | 276.30µs                        |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| LLVM_module_codegen                                                     | 147.60ms  | 2.599           | 2.60s    | 124        | 0.00ns                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| mir_built                                                               | 57.96ms   | 1.020           | 127.40ms | 969        | 9.19ms                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| LLVM_module_optimize                                                    | 50.75ms   | 0.893           | 50.75ms  | 124        | 0.00ns                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| codegen_select_candidate                                                | 49.52ms   | 0.872           | 54.98ms  | 2140       | 768.90µs                        |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| expand_crate                                                            | 48.77ms   | 0.859           | 56.16ms  | 1          | 0.00ns                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| items_of_instance                                                       | 38.66ms   | 0.681           | 287.07ms | 4439       | 3.65ms                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| fn_abi_of_instance                                                      | 37.04ms   | 0.652           | 71.64ms  | 6592       | 5.02ms                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
| evaluate_obligation                                                     | 35.74ms   | 0.629           | 47.79ms  | 4950       | 1.23ms                          |
+-------------------------------------------------------------------------+-----------+-----------------+----------+------------+---------------------------------+
```