@stefnotch 2025-09-13
I could spend some time asking the half-rs maintainer to tweak the libm thing, but I'm not sure if it's a good use of anyone's time.
We're getting closer and closer to the stabilization of a native f16 type https://github.com/rust-lang/rust/issues/116909 , which would let us replace half-rs with a blessed std implementation.

What do you think?